requirements.txt
streamlit
requests
deep-translator
openai-whisper
numpy
librosa